flow_envio_resultados_vd_browserless/requirements.txt
playwright==1.57.0
playwright-stealth
python-dotenv
requests
pyotp
